Rename rangeShape, NewRange, & RangeProp to Range
